fix some tslint errors and fix case-sensitive problems
