+++ package.json
@@ -44,6 +44,7 @@
         "@types/react-router-dom": "^4.3.1",
         "@types/serve-static": "^1.13.2",
         "@types/uuid": "^3.4.4",
+        "case-sensitive-paths-webpack-plugin": "^2.2.0",
         "enzyme": "^3.6.0",
         "enzyme-adapter-react-16": "^1.5.0",
         "express": "^4.16.4",
