+++ src/tests/unit/tests/scanner/message-decorator.test.ts
@@ -9,7 +9,6 @@ import { CustomRulesConfigurationStub } from './custom-rules-configuration-stub'
 
 describe('MessageDecorator', () => {
     let axeResultStub: AxeRule;
-    let passedAxeResultStub: AxeRule;
     let checkMessageTransformerMock: IMock<CheckMessageTransformer>;
     let configuration: RuleConfiguration[];
