+++ tslint.build-enforced.json
@@ -8,9 +8,11 @@
         "curly": true,
         "eofline": true,
         "forin": true,
+        "indent": [true, "spaces"],
         "interface-name": [true, "never-prefix"],
         "jsdoc-format": true,
         "label-position": true,
+        "max-classes-per-file": [true, 2],
         "member-access": true,
         "new-parens": true,
         "no-angle-bracket-type-assertion": true,
@@ -35,7 +37,9 @@
         "no-this-assignment": true,
         "no-trailing-whitespace": true,
         "no-unused-expression": true,
+        "no-var-keyword": true,
         "no-var-requires": true,
+        "prefer-const": [true, { "destructuring": "all" }],
         "one-line": [true, "check-open-brace", "check-whitespace"],
         "one-variable-per-declaration": [true, "ignore-for-loop"],
         "ordered-imports": true,
